@@ -3,33 +3,48 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@babel/runtime": "^7.9.2",
-    "@material-ui/core": "^4.9.1",
-    "@material-ui/icons": "^4.9.1",
     "antd": "^3.2.2",
     "axios": "^0.19.2",
-    "bootstrap": "^4.1.1",
-    "css-loader": "0.28.4",
-    "react": "^16.5.2",
-    "react-cookie": "^4.0.3",
-    "react-dom": "^16.5.2",
-    "react-minimal-pie-chart": "^6.0.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "1.1.5",
-    "react-youtube": "^7.11.2",
+    "cra-template": "1.1.3",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-redux": "^7.2.8",
+    "react-router-dom": "^6.0.0",
+    "react-scripts": "5.0.0",
     "redux": "^4.0.5",
     "universal-cookie": "^4.0.3"
   },
   "scripts": {
-    "start": "react-app-rewired start",
-    "build": "react-app-rewired build",
-    "test": "react-app-rewired test --env=jsdom",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "devDependencies": {
-    "babel-plugin-import": "^1.6.5",
-    "react-app-rewire-less": "^2.1.0",
-    "react-app-rewired": "^1.4.1",
-    "react-redux": "^7.1.3"
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.1",
+    "@mui/material": "^5.6.1",
+    "install": "^0.13.0",
+    "npm": "^8.7.0",
+    "web-vitals": "^2.1.4"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
