@@ -1,28 +1,28 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import './App.css';
 import {
+    Routes,
     Route,
-    withRouter,
-    Switch
-} from 'react-router-dom';
+    useNavigate,
+  } from "react-router-dom";
 
 import { getCurrentUser } from '../util/APIUtils';
 import {ACCESS_TOKEN} from '../constants';
-import Home from '../pages/home/Home';
+import Home from 'pages/home/Home';
 import Login from '../user/login/Login';
 import Signup from '../user/signup/Signup';
 import AppHeader from '../common/AppHeader';
 import NotFound from '../common/NotFound';
 import LoadingIndicator from '../common/LoadingIndicator';
-import {Layout, notification} from 'antd';
-import { createMuiTheme } from '@material-ui/core/styles';
-import {ThemeProvider} from "@material-ui/styles";
+import { createTheme } from '@mui/material/styles';
+import { ThemeProvider } from "@mui/material/styles";
 import PrivateRoute from "../common/PrivateRoute";
 import AdminRefresh from "../pages/admin/AdminRefresh";
-const {Content} = Layout;
+import ManagerCreateReservation from '../pages/createreservation/manager/ManagerCreateReservation';
+import ManagerViewReservations from '../pages/viewreservations/manager/ManagerViewReservations';
 
 //Formatting
-const theme = createMuiTheme({
+const theme = createTheme({
         palette: {
             primary: {
                 main: '#3773B0'
@@ -36,125 +36,86 @@ const theme = createMuiTheme({
     },
 );
 
-class App extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            currentUser: null, //Username of logged in
-            isAuthenticated: false, //Are they logged in
-            isLoading: false
-        };
-        
-        this.handleLogout = this.handleLogout.bind(this);
-        this.loadCurrentUser = this.loadCurrentUser.bind(this);
-        this.handleLogin = this.handleLogin.bind(this);
+export default function App(props) {
 
-        notification.config({
-            placement: 'topRight',
-            top: 70,
-            duration: 3,
-        });
-    }
+    const [currentUser, setCurrentUser] = useState(null);
+    const [isAuthenticated, setIsAuthenticated] = useState(false);
+    const [isLoading, setIsLoading] = useState(false);
+    const navigate = useNavigate();
 
-    loadCurrentUser() {
-        this.setState({
-            isLoading: true
-        });
+    useEffect(() => {
+        loadCurrentUser();
+    }, []);
+
+    const loadCurrentUser = () => {
+        setIsLoading(true);
 
         getCurrentUser()
             .then(response => {
-                this.setState({
-                    currentUser: response,
-                    isAuthenticated: true,
-                    isLoading: false
-                });
+                setCurrentUser(response);
+                setIsAuthenticated(true);
+                setIsLoading(false);
             }).catch(error => {
-            this.setState({
-                isLoading: false
+                setIsLoading(false);
             });
-        });
     }
 
-    componentDidMount() {
-        this.loadCurrentUser();
-    }
-
-    handleLogout(redirectTo = "/", notificationType = "success", description = "You're successfully logged out.") {
+    const handleLogout = (redirectTo = "/") => {
         localStorage.removeItem(ACCESS_TOKEN);
+        setCurrentUser(null);
+        setIsAuthenticated(false);
+        navigate(redirectTo);
 
-        this.setState({
-            currentUser: null,
-            isAuthenticated: false
-        });
-
-        this.props.history.push(redirectTo);
-
-        notification[notificationType]({
-            message: 'Hotel Management System',
-            description: description,
-        });
+        // notification[notificationType]({
+        //     message: 'Hotel Management System',
+        //     description: "You're successfully logged out.",
+        // });
     }
 
-    handleLogin(redirectTo = "/") {
-        notification.success({
-            message: 'Hotel Management System',
-            description: "You're successfully logged in.",
-        });
-        this.setState({
-            isLoading: true
-        });
+    const handleLogin = (redirectTo = "/") => {
+        // notification.success({
+        //     message: 'Hotel Management System',
+        //     description: "You're successfully logged in.",
+        // });
+        setIsLoading(true);
 
-        getCurrentUser().then(response => {
-            console.log(response);
-            this.setState({
-                currentUser: response,
-                isAuthenticated: true,
-                isLoading: false
+        getCurrentUser()
+            .then(response => {
+                setCurrentUser(response);
+                setIsAuthenticated(true);
+                setIsLoading(false);
+            }).catch(error => {
+                setIsLoading(false);
             });
-        }).catch(error => {
-            this.setState({
-                isLoading: false
-            });
-        });
 
-        this.props.history.push(redirectTo);
+        navigate(redirectTo);
     }
 
-    setSchool(school) {
-        this.setState({clickedSchool: school});
+    if (isLoading) {
+        return <LoadingIndicator/>
     }
 
-    render() {
-        if (this.state.isLoading) {
-            return <LoadingIndicator/>
-        }
-        return (
-            <ThemeProvider theme={theme}>
-                <Layout className="app-container">
-                <AppHeader isAuthenticated={this.state.isAuthenticated}
-                           currentUser={this.state.currentUser}
-                           onLogout={this.handleLogout}
-                           onSubChange={this.onSubChange}/>
-
-                <Content className="app-content">
-                    <div className="container">
-                        <Switch>
-                            <Route exact path="/" render={(props) => <Home isAuthenticated={this.state.isAuthenticated}  {...props} />}/>
-                            <Route path="/signup" component={Signup}/>
-                            <Route path="/login" render={(props) => <Login onLogin={this.handleLogin} {...props} />}/>
-
-                            <PrivateRoute exact path="/admin/refresh">
-                                <AdminRefresh/>
-                            </PrivateRoute>
-
-                            <Route component={NotFound}/>
-                        </Switch>
-                    </div>
-                </Content>
-            </Layout>
-            </ThemeProvider>
-        );
-    }
+    return (
+        <ThemeProvider theme={theme}>
+            <AppHeader isAuthenticated={isAuthenticated}
+                        currentUser={currentUser}
+                        onLogout={handleLogout} />
+            <div className="app-content">
+                <div className="container">
+                    <Routes>
+                        <Route exact path="/" element={<Home user={currentUser} isAuthenticated={isAuthenticated}  {...props} />}/>
+                        <Route path="signup" element={<Signup />}/>
+                        <Route path="login" element={<Login onLogin={handleLogin} {...props} />}/>
+                        <Route path="error" element={<NotFound />}/>
+                        <Route path="admin" element={<PrivateRoute />}>
+                            <Route path="createreservation" element={<ManagerCreateReservation />}/>
+                            <Route path="viewreservations" element={<ManagerViewReservations />}/>
+                            <Route path="refresh" element={<AdminRefresh />}/>
+                        </Route>
+                    </Routes>
+                </div>
+            </div>
+        </ThemeProvider>
+    );
+    
 }
-
-export default withRouter(App);
