@@ -1,254 +1,107 @@
-import React, { Component } from 'react';
-import { signup, checkUsernameAvailability } from '../../util/APIUtils';
+import * as React from 'react';
+import {useState} from 'react';
+import { useNavigate } from 'react-router-dom';
+import Button from '@mui/material/Button';
+import CssBaseline from '@mui/material/CssBaseline';
+import TextField from '@mui/material/TextField';
+import Link from '@mui/material/Link';
+import Grid from '@mui/material/Grid';
+import Box from '@mui/material/Box';
+import Typography from '@mui/material/Typography';
+import Container from '@mui/material/Container';
+import { createTheme, ThemeProvider } from '@mui/material/styles';
+import { signup, checkUsernameAvailability } from 'util/APIUtils';
 import './Signup.css';
-import { Link } from 'react-router-dom';
 import { 
     USERNAME_MIN_LENGTH, USERNAME_MAX_LENGTH,
-    PASSWORD_MIN_LENGTH, PASSWORD_MAX_LENGTH
-} from '../../constants';
+    PASSWORD_MIN_LENGTH, PASSWORD_MAX_LENGTH,
+    ACCESS_TOKEN
+} from 'constants';
 
-import { Form, Input, Button, notification } from 'antd';
-const FormItem = Form.Item;
+import Alert from '@mui/material/Alert';
+import AlertTitle from '@mui/material/AlertTitle';
+import Stack from '@mui/material/Stack';
 
-class Signup extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            username: {
-                value: ''
-            },
-            password: {
-                value: ''
-            },
-            firstname: {
-                value: ''
-            },
-            lastname: {
-                value: ''
-            },
-            age: {
-                value: ''
-            },
-            email: {
-                value: ''
-            },
-            phonenumber: {
-                value: ''
-            },
-            phase: 1
-        };
+const theme = createTheme();
 
-        this.handleNext = this.handleNext.bind(this);
-        this.handleInputChange = this.handleInputChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
-        this.validateUsernameAvailability = this.validateUsernameAvailability.bind(this);
-        this.isFormInvalid = this.isFormInvalid.bind(this);
+export default function SignUp(props) {
+
+    const navigate = useNavigate();
+
+    let [alert, setAlert] = useState(null);
+    let [username, setUsername] = useState({
+        value: ''
+    });
+    let [password, setPassword] = useState({
+        value: ''
+    });
+    let [firstname, setFirstname] = useState({
+        value: ''
+    });
+    let [lastname, setLastname] = useState({
+        value: ''
+    });
+    let [age, setAge] = useState({
+        value: ''
+    });
+    let [email, setEmail] = useState({
+        value: ''
+    });
+    let [phonenumber, setPhonenumber] = useState({
+        value: ''
+    });
+    let [phase, setPhase] = useState(1);
+
+    const setters = {
+        username: setUsername,
+        password: setPassword,
+        firstname: setFirstname,
+        lastname: setLastname,
+        age: setAge,
+        email: setEmail,
+        phonenumber: setPhonenumber
     }
 
-    handleInputChange(event, validationFun) {
+    const handleInputChange = (event, validationFun) => {
         const target = event.target;
         const inputName = target.name;        
         const inputValue = target.value;
 
-        this.setState({
-            [inputName] : {
-                value: inputValue,
-                ...validationFun(inputValue)
-            }
+        const setter = setters[inputName];
+        setter({
+            value: inputValue,
+            ...validationFun(inputValue)
         });
     }
 
-    handleSubmit() {
-
-        const signupRequest = {
-            username: this.state.username.value,
-            password: this.state.password.value,
-            firstName: this.state.firstname.value,
-            lastName: this.state.lastname.value,
-            age: this.state.age.value,
-            email: this.state.email.value,
-            phoneNumber: this.state.phonenumber.value,
-        };
-        
-        signup(signupRequest)
-            .then(response => {
-                notification.success({
-                    message: 'Hotel Management System',
-                    description: "Thank you! You're successfully registered. Please Login to continue!",
-                });
-                this.props.history.push("/login");
-            }).catch(error => {
-            notification.error({
-                message: 'Hotel Management System',
-                description: error.message || 'Sorry! Something went wrong. Please try again!'
-            });
-        });
+    const debug = () => {
+        console.log(username);
+        console.log(password);
+        console.log(firstname);
+        console.log(lastname);
+        console.log(age);
+        console.log(email);
+        console.log(phonenumber);
     }
 
-    isFormInvalid() {
+    const isFormInvalid = () => {
         return !(
-            this.state.username.validateStatus === 'success' &&
-            this.state.password.validateStatus === 'success' &&
-            this.state.firstname.validateStatus === 'success' &&
-            this.state.lastname.validateStatus === 'success' &&
-            this.state.age.validateStatus === 'success' &&
-            this.state.email.validateStatus === 'success' &&
-            this.state.phonenumber.validateStatus === 'success'
+            username.validateStatus === 'success' &&
+            password.validateStatus === 'success' &&
+            firstname.validateStatus === 'success' &&
+            lastname.validateStatus === 'success' &&
+            age.validateStatus === 'success' &&
+            email.validateStatus === 'success' &&
+            phonenumber.validateStatus === 'success'
         );
     };
 
-    handleNext() {
-        this.setState({
-            phase: this.state.phase + 1
-        });
-    }
-
-    render() {
-        
-        // First name, last name, email, age, phonenumber
-        const formPhaseOne = (
-            <Form className="signup-form">
-                <FormItem label="First Name"
-                    hasFeedback
-                    validateStatus={this.state.firstname.validateStatus}
-                    help={this.state.firstname.errorMsg}>
-                    <Input
-                        size="large"
-                        name="firstname"
-                        autoComplete="off"
-                        placeholder="First Name"
-                        value={this.state.firstname.value}
-                        onChange={(event) => this.handleInputChange(event, this.validatePassthrough)} />
-                </FormItem>
-                <FormItem
-                    label="Last Name"
-                    validateStatus={this.state.lastname.validateStatus}
-                    help={this.state.lastname.errorMsg}>
-                    <Input
-                        size="large"
-                        name="lastname"
-                        autoComplete="off"
-                        placeholder="Last Name"
-                        value={this.state.lastname.value}
-                        onChange={(event) => this.handleInputChange(event, this.validatePassthrough)} />
-                </FormItem>
-                <FormItem
-                    label="Age"
-                    validateStatus={this.state.age.validateStatus}
-                    help={this.state.age.errorMsg}>
-                    <Input
-                        size="large"
-                        name="age"
-                        autoComplete="off"
-                        placeholder="Age"
-                        value={this.state.age.value}
-                        onChange={(event) => this.handleInputChange(event, this.validatePassthrough)} />
-                </FormItem>
-                <FormItem
-                    label="Email"
-                    validateStatus={this.state.email.validateStatus}
-                    help={this.state.email.errorMsg}>
-                    <Input
-                        size="large"
-                        name="email"
-                        autoComplete="off"
-                        placeholder="Email"
-                        value={this.state.email.value}
-                        onChange={(event) => this.handleInputChange(event, this.validatePassthrough)} />
-                </FormItem>
-                <FormItem
-                    label="Phone Number"
-                    validateStatus={this.state.phonenumber.validateStatus}
-                    help={this.state.phonenumber.errorMsg}>
-                    <Input
-                        size="large"
-                        name="phonenumber"
-                        autoComplete="off"
-                        placeholder="Phone Number"
-                        value={this.state.phonenumber.value}
-                        onChange={(event) => this.handleInputChange(event, this.validatePassthrough)} />
-                </FormItem>
-                <FormItem>
-                    <Button type="primary"
-                        htmlType="submit"
-                        size="large"
-                        className="signup-form-button"
-                        disabled={this.isFormInvalid()}
-                        onClick={this.handleNext}>Next
-                    </Button>
-                    Already registed? <Link to="/login">Login now!</Link>
-                </FormItem>
-            </Form>
-        );
-
-        const formPhaseTwo = (
-            <Form className="signup-form">
-                <FormItem label="Username"
-                    hasFeedback
-                    validateStatus={this.state.username.validateStatus}
-                    help={this.state.username.errorMsg}>
-                    <Input
-                        size="large"
-                        name="username"
-                        autoComplete="off"
-                        placeholder="A unique username"
-                        value={this.state.username.value}
-                        onChange={(event) => this.handleInputChange(event, this.validateUsername)} />
-                </FormItem>
-                <FormItem
-                    label="Password"
-                    validateStatus={this.state.password.validateStatus}
-                    help={this.state.password.errorMsg}>
-                    <Input
-                        size="large"
-                        name="password"
-                        type="password"
-                        autoComplete="off"
-                        placeholder="A password between 6 to 20 characters"
-                        value={this.state.password.value}
-                        onChange={(event) => this.handleInputChange(event, this.validatePassword)} />
-                </FormItem>
-                <FormItem>
-                    <Button type="primary"
-                        htmlType="submit"
-                        size="large"
-                        className="signup-form-button"
-                        disabled={this.isFormInvalid()}
-                        onClick={this.handleSubmit}>Submit
-                    </Button>
-                    Already registed? <Link to="/login">Login now!</Link>
-                </FormItem>
-            </Form>
-        );
-
-        let currentForm = null;
-        switch(this.state.phase) {
-            case 1: 
-                currentForm = formPhaseOne; 
-                break;
-            case 2: 
-                currentForm = formPhaseTwo; 
-                break;
-            default:
-                console.log('Trying to render phase ' + this.state.phase + ' for signup form.');
-                break;
-        }
-
-        return (
-            <div>
-                <div className="signup-container">
-                    <h1 className="page-title">Sign Up</h1>
-                    <div className="signup-content">
-                        {currentForm}
-                    </div>
-                </div>
-            </div>
-        );
+    const handleNext = () => {
+        setPhase(phase + 1);
     }
 
     // Validation Functions
-
-    validateUsername = (username) => {
+    const validateUsername = () => {
         if(username.length < USERNAME_MIN_LENGTH) {
             return {
                 validateStatus: 'error',
@@ -261,19 +114,20 @@ class Signup extends Component {
             }
         } else {
             return {
-                validateStatus: null,
+                validateStatus: 'success',
                 errorMsg: null
             }
         }
     };
 
-    validateUsernameAvailability() {
+    const validateUsernameAvailability = () => {
+
         // First check for client side errors in username
-        const usernameValue = this.state.username.value;
-        const usernameValidation = this.validateUsername(usernameValue);
+        const usernameValue = username.value;
+        const usernameValidation = validateUsername(usernameValue);
 
         if(usernameValidation.validateStatus === 'error') {
-            this.setState({
+            setUsername({
                 username: {
                     value: usernameValue,
                     ...usernameValidation
@@ -282,46 +136,38 @@ class Signup extends Component {
             return;
         }
 
-        this.setState({
-            username: {
-                value: usernameValue,
-                validateStatus: 'validating',
-                errorMsg: null
-            }
+        setUsername({
+            value: usernameValue,
+            validateStatus: 'validating',
+            errorMsg: null
         });
 
         checkUsernameAvailability(usernameValue)
         .then(response => {
             if(response.available) {
-                this.setState({
-                    username: {
-                        value: usernameValue,
-                        validateStatus: 'success',
-                        errorMsg: null
-                    }
-                });
-            } else {
-                this.setState({
-                    username: {
-                        value: usernameValue,
-                        validateStatus: 'error',
-                        errorMsg: 'This username is already taken'
-                    }
-                });
-            }
-        }).catch(error => {
-            // Marking validateStatus as success, Form will be recchecked at server
-            this.setState({
-                username: {
+                setUsername({
                     value: usernameValue,
                     validateStatus: 'success',
                     errorMsg: null
-                }
+                });
+            } else {
+                
+                setUsername({
+                    value: usernameValue,
+                    validateStatus: 'error',
+                    errorMsg: 'This username is already taken'
+                });
+            }
+        }).catch(error => {
+            setUsername({
+                value: usernameValue,
+                validateStatus: 'success',
+                errorMsg: null
             });
         });
     }
 
-    validatePassword = (password) => {
+    const validatePassword = (password) => {
         if(password.length < PASSWORD_MIN_LENGTH) {
             return {
                 validateStatus: 'error',
@@ -340,13 +186,211 @@ class Signup extends Component {
         }
     }
 
-    validatePassthrough = (field) => {
+    const validatePassthrough = () => {
         return {
             validateStatus: 'success',
             errorMsg: null,
         };
     }
 
-}
+    const handleSubmit = () => {
+        
+        debug();
 
-export default Signup;
+        if(isFormInvalid()) {
+            setAlert(
+                <Alert severity="error">
+                    <AlertTitle>Error</AlertTitle>
+                    Please try again!
+                </Alert>
+            );
+        }
+
+        const signupRequest = {
+            username: username.value,
+            password: password.value,
+            firstName: firstname.value,
+            lastName: lastname.value,
+            age:age.value,
+            email: email.value,
+            phoneNumber: phonenumber.value,
+        };
+
+        console.log('signupRequest', signupRequest);
+
+        signup(signupRequest)
+            .then(response => {
+                // notification.success({
+                //     message: 'Hotel Management System',
+                //     description: "Thank you! You're successfully registered. Please Login to continue!",
+                // });
+                navigate("/login");
+            }).catch(error => {
+                setAlert(
+                    <Alert severity="error">
+                        <AlertTitle>Error</AlertTitle>
+                        Sorry! Something went wrong. Please try again!
+                    </Alert>
+                );
+            });
+    };
+
+    // First name, last name, email, age, phonenumber
+    const formPhaseOne = (
+        <Box component="form" onSubmit={handleSubmit} noValidate sx={{ mt: 1 }}>
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="firstname"
+                label="First Name"
+                type="firstname"
+                id="firstname"
+                autoComplete="firstname"
+                value={firstname.value}
+                onChange={(event) => handleInputChange(event, validatePassthrough)} 
+            />
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="lastname"
+                label="Last Name"
+                type="lastname"
+                id="lastname"
+                autoComplete="lastname"
+                value={lastname.value}
+                onChange={(event) => handleInputChange(event, validatePassthrough)} 
+            />
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="age"
+                label="Age"
+                type="age"
+                id="age"
+                autoComplete="age"
+                value={age.value}
+                onChange={(event) => handleInputChange(event, validatePassthrough)} 
+            />
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="email"
+                label="Email"
+                type="email"
+                id="email"
+                autoComplete="email"
+                value={email.value}
+                onChange={(event) => handleInputChange(event, validatePassthrough)} 
+            />
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                id="phonenumber"
+                label="Phone Number"
+                name="phonenumber"
+                autoComplete="phonenumber"
+                value={phonenumber.value}
+                onChange={(event) => handleInputChange(event, validatePassthrough)} 
+            />
+            <Button
+                onClick={handleNext}
+                fullWidth
+                variant="contained"
+                sx={{ mt: 3, mb: 2 }}
+            >
+                Next
+            </Button>
+            <Grid container>
+                <Grid item>
+                    <Link href="/login" variant="body2">
+                    {"Already have an account? Sign In"}
+                    </Link>
+                </Grid>
+            </Grid>
+        </Box>
+    );
+
+    const formPhaseTwo = (
+        <Box component="form" onSubmit={handleSubmit} noValidate sx={{ mt: 1 }}>
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="username"
+                label="Username"
+                type="username"
+                id="username"
+                autoComplete="username"
+                value={username.value}
+                onChange={(event) => handleInputChange(event, validateUsername)} 
+            />
+            <TextField
+                margin="normal"
+                required
+                fullWidth
+                name="password"
+                label="Password"
+                type="password"
+                id="password"
+                autoComplete="password"
+                value={password.value}
+                onChange={(event) => handleInputChange(event, validatePassword)} 
+            />
+            <Button
+                onClick={handleSubmit}
+                fullWidth
+                variant="contained"
+                sx={{ mt: 3, mb: 2 }}
+            >
+                Submit
+            </Button>
+            <Grid container>
+                <Grid item>
+                    <Link href="/login" variant="body2">
+                    {"Already have an account? Sign In"}
+                    </Link>
+                </Grid>
+            </Grid>
+        </Box>
+    );
+
+    let currentForm = null;
+    switch(phase) {
+        case 1: 
+            currentForm = formPhaseOne; 
+            break;
+        case 2: 
+            currentForm = formPhaseTwo; 
+            break;
+        default:
+            console.log('Trying to render phase ' + phase + ' for signup form.');
+            break;
+    }
+  
+    return (
+      <ThemeProvider theme={theme}>
+        <Container component="main" maxWidth="xs">
+          <CssBaseline />
+          <Stack sx={{ width: '100%' }} spacing={2}>{alert}</Stack>
+          <Box
+            sx={{
+              marginTop: 8,
+              display: 'flex',
+              flexDirection: 'column',
+              alignItems: 'center',
+            }}
+          >
+            <Typography component="h1" variant="h5">
+              Sign up
+            </Typography>
+            {currentForm}
+          </Box>
+        </Container>
+      </ThemeProvider>
+    );
+  }
